feat: add APScheduler as a systemd daemon
This commit introduces the ability to run APScheduler as a systemd daemon.
It includes:
- A guide on how to configure APScheduler as a systemd daemon.
- A management command to start the scheduler.
- A check to ensure the scheduler is only started when the
 DJANGO_RUN_SCHEDULER environment variable is set to "1".

The changes allow the scheduler to run independently of the web
server, improving reliability and scalability. The scheduler is now
started only when the DJANGO_RUN_SCHEDULER environment variable is set
to "1". This allows the scheduler to be started and stopped
independently of the web server.